+++ temba/api/v2/views.py
@@ -351,7 +351,6 @@ class ListAPIMixin(mixins.ListModelMixin):
     Mixin for any endpoint which returns a list of objects from a GET request
     """
     exclusive_params = ()
-    required_params = ()
 
     def get(self, request, *args, **kwargs):
         return self.list(request, *args, **kwargs)
@@ -369,11 +368,6 @@ class ListAPIMixin(mixins.ListModelMixin):
         # check user hasn't provided values for more than one of any exclusive params
         if sum([(1 if params.get(p) else 0) for p in self.exclusive_params]) > 1:
             raise InvalidQueryError("You may only specify one of the %s parameters" % ", ".join(self.exclusive_params))
-
-        # check that any required params are included
-        if self.required_params:
-            if sum([(1 if params.get(p) else 0) for p in self.required_params]) != 1:
-                raise InvalidQueryError("You must specify one of the %s parameters" % ", ".join(self.required_params))
 
     def filter_before_after(self, queryset, field):
         """
@@ -2186,6 +2180,8 @@ class MessagesEndpoint(ListAPIMixin, BaseAPIView):
     You can also filter by `folder` where folder is one of `inbox`, `flows`, `archived`, `outbox`, `incoming`, `failed` or `sent`.
     Note that you cannot filter by more than one of `contact`, `folder`, `label` or `broadcast` at the same time.
 
+    Without any parameters this endpoint will return all incoming and outgoing messages ordered by creation date.
+
     The sort order for all folders save for `incoming` is the message creation date. For the `incoming` folder (which
     includes all incoming messages, regardless of visibility or type) messages are sorted by last modified date. This
     allows clients to poll for updates to message labels and visibility changes.
@@ -2235,7 +2231,6 @@ class MessagesEndpoint(ListAPIMixin, BaseAPIView):
     serializer_class = MsgReadSerializer
     pagination_class = Pagination
     exclusive_params = ('contact', 'folder', 'label', 'broadcast')
-    required_params = ('contact', 'folder', 'label', 'broadcast', 'id')
     throttle_scope = 'v2.messages'
 
     FOLDER_FILTERS = {'inbox': SystemLabel.TYPE_INBOX,